% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_node_names_to_edges.R
\name{add_names}
\alias{add_names}
\title{Add note identifiers to edge list}
\usage{
add_names(G, id)
}
\arguments{
\item{G}{Graph object}

\item{id}{The new ID variable}
}
\value{

}
\description{
Add note identifiers to edge list
}
