% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_node_attribute.R
\name{add_node_attribute}
\alias{add_node_attribute}
\title{Add note identifiers to edge list}
\usage{
add_node_attribute(G, id)
}
\arguments{
\item{G}{Graph object}

\item{id}{The new ID variable}
}
\value{

}
\description{
Add note identifiers to edge list
}
