% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-firm-hhi.R
\name{calculate_firm_hhi}
\alias{calculate_firm_hhi}
\title{Calculate firm-level HHI}
\usage{
calculate_firm_hhi(A, markets)
}
\arguments{
\item{A}{Bipartite adjacency matrix}

\item{markets}{Named vector mapping each firm to its market}
}
\value{

}
\description{
Calculate firm-level HHI
}
