% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-geographic-market-hhi-from-firm-hhi.R
\name{calculate_geographic_market_hhi_from_firm_hhi}
\alias{calculate_geographic_market_hhi_from_firm_hhi}
\title{Construct a geographic HHI measure from firm-level HHI measures}
\usage{
calculate_geographic_market_hhi_from_firm_hhi(A, markets)
}
\arguments{
\item{A}{Bipartite adjacency matrix}

\item{markets}{Named vector of geographic markets}
}
\value{

}
\description{
Construct a geographic HHI measure from firm-level HHI measures
}
