% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-outflow-hhi.R
\name{calculate_outflow_hhi}
\alias{calculate_outflow_hhi}
\title{Calculate outflow HHI measures from a bipartite adjacency matrix}
\usage{
calculate_outflow_hhi(A, markets)
}
\arguments{
\item{A}{The bipartite adjacency matrix}

\item{markets}{Named vector mapping each micromarket to its market}
}
\value{

}
\description{
Calculate outflow HHI measures from a bipartite adjacency matrix
}
