% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-wtp.R
\name{calculate_wtp}
\alias{calculate_wtp}
\title{Calculate Willingness-to-Pay}
\usage{
calculate_wtp(A)
}
\arguments{
\item{A}{Bipartite adjacency matrix}
}
\value{

}
\description{
Calculate Willingness-to-Pay
}
