% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_graph.R
\name{collapse_graph}
\alias{collapse_graph}
\title{Collapse a graph network by an identifying node-level group variable}
\usage{
collapse_graph(G, x)
}
\arguments{
\item{G}{The graph network}

\item{x}{The group variable name (node attribute)}
}
\value{

}
\description{
Collapse a graph network by an identifying node-level group variable
}
