% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_adjacency.R
\name{convert_to_adjacency}
\alias{convert_to_adjacency}
\title{Convert a graph network to a weighted adjacency matrix}
\usage{
convert_to_adjacency(G, id = name)
}
\arguments{
\item{G}{The graph network}

\item{id}{The ID variable name}
}
\value{

}
\description{
Convert a graph network to a weighted adjacency matrix
}
