% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-example-adjacency.r
\name{create_example_adjacency}
\alias{create_example_adjacency}
\title{Create an example adjacency matrix}
\usage{
create_example_adjacency()
}
\value{

}
\description{
Create an example adjacency matrix
}
