% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-unipartite-adjacency.R
\name{create_unipartite_adjacency}
\alias{create_unipartite_adjacency}
\title{Create a unipartite adjacency matrix}
\usage{
create_unipartite_adjacency(A, MARGIN)
}
\arguments{
\item{A}{Bipartite adjacency matrix}

\item{MARGIN}{The margin (row=1 or column=2) over which to define the unipartite projection.}
}
\value{

}
\description{
Create a unipartite adjacency matrix
}
