% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-markets.R
\name{detect_markets}
\alias{detect_markets}
\title{Title}
\usage{
detect_markets(A, MARGIN)
}
\arguments{
\item{A}{Bipartite adjacency matrix}

\item{MARGIN}{The margin (row=1 or column=2) over which to define markets.}
}
\value{

}
\description{
Title
}
