% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ssnip.R
\name{fit_ssnip}
\alias{fit_ssnip}
\title{Fit a SSSNIP test based on a clustering of firm-specific markets.}
\usage{
fit_ssnip(A, target)
}
\arguments{
\item{A}{The input adjacency matrix}

\item{target}{The target hospital}
}
\value{

}
\description{
Fit a SSSNIP test based on a clustering of firm-specific markets.
}
