% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_hosp_A}
\alias{geo_hosp_A}
\title{Example bipartite adjanccy matrix}
\format{
A matrix with 12 rows and 10 columns
\describe{
\item{rows}{Geographic micromarkets}
\item{}{Hospitals}
...
}
}
\usage{
geo_hosp_A
}
\description{
A matrix summarizing aggregate demand within a bipartite adjacency matrix.
}
\keyword{datasets}
