% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bipartite-graph.r
\name{get_bipartite_graph}
\alias{get_bipartite_graph}
\title{Construct a bipartite graph network from a bipartite adjacency matrix}
\usage{
get_bipartite_graph(A, types = c("geography", "hospital"))
}
\arguments{
\item{A}{The bipartite adjacency matrix}

\item{types}{Names for the bipartite nodes c(row_name, column_name)}
}
\value{
A bipartite graph object
}
\description{
Construct a bipartite graph network from a bipartite adjacency matrix
}
