% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-distance-members.R
\name{get_distance_members}
\alias{get_distance_members}
\title{Identify points within a great circle distance}
\usage{
get_distance_members(df, id, coords, dist, unit = c("miles", "km"))
}
\arguments{
\item{df}{The input data frame}

\item{id}{The identifying variable}

\item{coords}{Nested coordinate object with x,y coordinates for the point}

\item{dist}{The maximum distance allowed}

\item{unit}{The unit (miles or km)}
}
\value{

}
\description{
Identify points within a great circle distance
}
