% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-greatcircle-distances.R
\name{get_greatcircle_distances}
\alias{get_greatcircle_distances}
\title{Get great circle distances between coordinates in a data file}
\usage{
get_greatcircle_distances(df, id, coords)
}
\arguments{
\item{df}{The input data file}

\item{id}{The ID variable}

\item{coords}{A nested data frame with x,y coordinates of the point}
}
\value{

}
\description{
Get great circle distances between coordinates in a data file
}
