% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso30}
\alias{iso30}
\title{Example isochrone}
\format{
A list object with 5 elements, each of which is an isochrone sf geometry around a point.
\describe{
\item{elements}{60-minute drive-time isochrone}
...
}
}
\usage{
data(iso30)
}
\description{
An sf list object with isochrones (60min) for four x,y locations.
}
\keyword{datasets}
