% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-merger.R
\name{simulate_merger}
\alias{simulate_merger}
\title{Simulate a merger of nodes in the bipartite adjacency matrix}
\usage{
simulate_merger(A, .x, .y)
}
\arguments{
\item{A}{The bipartite adjacency matrix}

\item{.x}{The absorbing firm}

\item{.y}{The absorbee.}
}
\value{

}
\description{
Simulate a merger of nodes in the bipartite adjacency matrix
}
