% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnip_.R
\name{ssnip_}
\alias{ssnip_}
\title{Fit an SSNIP test (note: invoked by fit_ssnip())}
\usage{
ssnip_(df, A, target)
}
\arguments{
\item{df}{Data frame}

\item{A}{Adjacency matrix}

\item{target}{The target hospital for SSNIP test.}
}
\value{

}
\description{
Fit an SSNIP test (note: invoked by fit_ssnip())
}
